orders/requirements.txt
alembic==1.0.11
aniso8601==7.0.0
Click==7.0
Flask==1.1.1
Flask-JWT==0.3.2
flask-marshmallow==0.10.1
Flask-Migrate==2.5.2
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.4.0
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.1.0
MarkupSafe==1.1.1
marshmallow==2.19.5
psycopg2==2.8.3
PyJWT==1.4.2
python-dateutil==2.8.0
python-editor==1.0.4
pytz==2019.2
six==1.12.0
SQLAlchemy==1.3.6
Werkzeug==0.15.5
